pyproject.toml
[project]
name = "pack-mm"
version = "0.0.14"
description = "packing materials and molecules in boxes using for machine learnt interatomic potentials"
authors = [
    { name = "Alin M. Elena" },
]
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Development Status :: 3 - Alpha",
]
readme = "README.md"

dependencies = [
    "janus-core>=0.7.2",
    "typer<1.0.0,>=0.12.5",
    "typer-config<2.0.0,>=1.4.2",
]


[project.scripts]
packmm = "pack_mm.cli.packmm:app"

[project.urls]
Repository = "https://github.com/ddmms/pack-mm/"
Documentation = "https://ddmms.github.io/pack-mm/"

[dependency-groups]
dev = [
    "coverage[toml]<8.0.0,>=7.4.1",
    "ipykernel>=6.29.5",
    "pgtest<2.0.0,>=1.3.2",
    "pytest<9.0,>=8.0",
    "pytest-cov<5.0.0,>=4.1.0",
    "tox-uv<2.0,>=1.16.1",
    "wheel<1.0,>=0.42",
]

docs = [
    "furo<2025.0.0,>=2024.1.29",
    "jupyter>=1.1.1",
    "markupsafe<2.1",
    "nbsphinx>=0.9.6",
    "numpydoc<2.0.0,>=1.6.0",
    "sphinx<9.0.0,>=8.0.2",
    "sphinxcontrib-contentui<1.0.0,>=0.2.5",
    "sphinxcontrib-details-directive<1.0,>=0.1",
    "sphinx-autodoc-typehints<3.0.0,>=2.5.0",
    "sphinx-collapse>=0.1.3",
    "sphinx-copybutton<1.0.0,>=0.5.2",
]

pre-commit = [
    "pre-commit<4.0.0,>=3.6.0",
    "ruff<1.0.0,>=0.9.3",
]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pytest.ini_options]
# Configuration for [pytest](https://docs.pytest.org)
python_files = "test_*.py"
addopts = '--cov-report xml'
pythonpath = ["."]

[tool.coverage.run]
# Configuration of [coverage.py](https://coverage.readthedocs.io)
# reporting which lines of your plugin are covered by tests
source=["pack_mm"]

[tool.ruff]
exclude = ["conf.py", "*ipynb"]
target-version = "py310"

[tool.ruff.lint]
# Ignore complexity
ignore = ["C901","B008"]
select = [
    # flake8-bugbear
    "B",
    # pylint
    "C", "R",
    # pydocstyle
    "D",
    # pycodestyle
    "E", "W",
    # Pyflakes
    "F", "FA",
    # pyupgrade
    "I",
    # pep8-naming
    "N",
    # isort
    "UP",
]

[tool.ruff.lint.isort]
force-sort-within-sections = true
required-imports = ["from __future__ import annotations"]

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.ruff.lint.pylint]
max-args = 10

[tool.ruff.lint.pyupgrade]
keep-runtime-typing = false

[tool.numpydoc_validation]
# report on all checks, except the below
checks = [
    "all",
    "EX01",
    "SA01",
    "ES01",
    "PR04", # Ignore no type (types come from signature as via sphinx_autodoc_typehints)
]
# Don't report on objects that match any of these regex
exclude = [
    ".__weakref__$",
    ".__repr__$",
]

[tool.uv]
default-groups = [
    "dev",
    "docs",
    "pre-commit",
]
